cloud functions/services_weather-updater-v1_1761227451.545000/requirements.txt
functions-framework
requests
pandas
sqlalchemy
pymysql
pytz
python-dotenv
beautifulsoup4
lat-lon-parser